+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Added rexbug
 ### Changed
 - Changed Command.Tl to Command.Translate
+- Refactored translate
 
 ## [1.16.0] - 2020-10-02
 ### Added